+++ local/docs/figures/utils/__init__.py
@@ -1,0 +1,5 @@
+from .save import save_fig
+from .style import set_docs_style
+
+# automatically set style
+set_docs_style()