+++ local/docs/figures/utils/style.py
@@ -1,0 +1,5 @@
+import matplotlib.pyplot as plt
+
+
+def set_docs_style():
+    plt.style.use("./local/docs/figures/docs.mplstyle")