+++ local/docs/utils/__init__.py
@@ -1,0 +1,1 @@
+from .kde import univariate_kde, univariate_kde_adaptive