+++ max_div/benchmarks/_problems/_problem_u4.py
@@ -48,10 +48,10 @@ class BenchmarkProblem_U4(BenchmarkProblem):
     def _create_problem_instance(cls, size: int, diversity_metric: DiversityMetric, **kwargs) -> MaxDivProblem:
         """
         We will generate vectors in d-dim. space as follows:
-          - [a, a, a, ..., a] + r * [x1, x2, ..., xd]
+          - a * ([1, 1, 1, ..., 1] + (r * [x1, x2, ..., xd]))
 
         - a is sampled uniformly in [0,1]
-        - x-values are sampled uniformly in the hyper-box [-1, +1]^d and then rescaled to have L2 norm = r
+        - xi-values are sampled uniformly in the hyper-box [-1, +1]^d and then rescaled to have L2 norm = 1
         - r = 0.1 * sqrt(d), such that the perceived angle of the cone from the origin remains constant as d increases
         """
 
@@ -69,8 +69,8 @@ class BenchmarkProblem_U4(BenchmarkProblem):
             x_vals = (random_data[i, 1:] * 2.0) - 1.0  # map [0,1] to [-1,+1]
             x_norm = np.linalg.norm(x_vals, ord=2)
             if x_norm > 0:
-                x_vals = (x_vals / x_norm) * r  # rescale to have L2 norm = r
-            vectors[i, :] = a + x_vals
+                x_vals = (x_vals / x_norm) * r  # rescale to have L2 norm = a*r
+            vectors[i, :] = a * (1 + x_vals)
 
         # step 3 - sort vectors by increasing L2 norm of rows & return problem instance
         vectors = sort_vectors(vectors)  # sort by increasing L2 norm of rows